+++ components/menu-dropdown/check-props.js
@@ -20,6 +20,14 @@ if (process.env.NODE_ENV !== 'production') {
 			'Please use isOpen instead. It provides a consistent prop that aligns with other componenents.'
 		);
 
+		deprecatedProperty(
+			COMPONENT,
+			props.offset,
+			'offset',
+			undefined,
+			'The manual setting of positional offset of dialog components has been deemed unreliable. Position logic has been re-written to deliver better and more reliable positioning. Please create an issue if you have an edge case not covered by the built-in logic.'
+		);
+
 		oneOfRequiredProperty(COMPONENT, {
 			options: props.options,
 			children: props.children,