+++ components/utilities/dialog/index.jsx
@@ -18,7 +18,11 @@ import Portal from './portal';
 import EventUtil from '../../../utilities/event';
 import KEYS from '../../../utilities/key-code';
 import DOMElementFocus from '../../../utilities/dom-element-focus';
-import { mapPropToPopperPlacement } from '../../../utilities/dialog-helpers';
+import {
+	getNubbinMargins,
+	getNubbinClassName,
+	mapPropToPopperPlacement,
+} from '../../../utilities/dialog-helpers';
 
 import { DIALOG } from '../../../utilities/constants';
 
@@ -95,15 +99,21 @@ const Dialog = createReactClass({
 		 */
 		containerProps: PropTypes.object,
 		/**
-		 * Sets the dialog width to the width of either 'target' (Menus attached to `input` typically follow this UX pattern), 'menu' or 'none.
+		 * Will show the nubbin pointing from the dialog to the reference element. Positioning and offsets will be handled.
 		 */
-		inheritWidthOf: PropTypes.oneOf(['target', 'menu', 'none']),
+		hasNubbin: PropTypes.bool,
 		/**
 		 * By default, dialogs will flip their alignment (such as bottom to top) if they extend beyond a boundary element such as a scrolling parent or a window/viewpoint. `hasStaticAlignment` disables this behavior and allows this component to extend beyond boundary elements.
 		 */
 		hasStaticAlignment: PropTypes.bool,
 		/**
-		 *  Offset adds pixels to the absolutely positioned dropdown menu in the format: ([vertical]px [horizontal]px). SHOULD BE OBJECT -----------
+		 * Sets the dialog width to the width of either 'target' (Menus attached to `input` typically follow this UX pattern), 'menu' or 'none.
+		 */
+		inheritWidthOf: PropTypes.oneOf(['target', 'menu', 'none']),
+		/**
+		 * DEPRECATED - do not add checkProp deprecation message at this level. It is handled at higher level components.
+		 * TODO - to be removed.
+		 * Offset adds pixels to the absolutely positioned dropdown menu in the format: ([vertical]px [horizontal]px). SHOULD BE OBJECT -----------
 		 */
 		offset: PropTypes.string,
 		/**
@@ -166,6 +176,7 @@ const Dialog = createReactClass({
 		 * An object of CSS styles that are applied to the immediate parent `div` of the contents. Use this instead of margin props.
 		 */
 		style: PropTypes.object,
+
 		/**
 		 * Sets which focus UX pattern to follow. For instance, popovers trap focus and must be exited to regain focus. Dropdowns and Tooltips never have focus.
 		 */
@@ -242,10 +253,21 @@ const Dialog = createReactClass({
 			};
 		}
 
-		const propOffsets = this.getPropOffsetsInPixels(this.props.offset);
 		const { position } = popperData.offsets.popper;
-		const left = `${popperData.offsets.popper.left + propOffsets.horizontal}px`;
-		const top = `${popperData.offsets.popper.top + propOffsets.vertical}px`;
+		const propOffsets = this.getPropOffsetsInPixels(this.props.offset);
+
+		// FIXME before merge - gotta rename from margin to offset
+		const nubbinOffsets = this.props.hasNubbin
+			? getNubbinMargins(this.state.popperData)
+			: { left: 0, top: 0 };
+
+		const left =
+			popperData.offsets.popper.left +
+			nubbinOffsets.left +
+			propOffsets.horizontal;
+		const top =
+			popperData.offsets.popper.top + nubbinOffsets.top + propOffsets.vertical;
+
 		// A Dropdown with overflowBoundaryElement position and 'align=right' uses max-width instead of inherited children width
 		const right = 'inherit';
 		return { ...popperData.style, left, top, right, position };
@@ -294,12 +316,10 @@ const Dialog = createReactClass({
 	},
 
 	handleOpen () {
-		const scopedElement = this.dialogContent;
-
-		if (this.props.variant === 'popover' && scopedElement) {
+		if (this.props.variant === 'popover' && this.dialogContent) {
 			DOMElementFocus.storeActiveElement();
 			DOMElementFocus.setupScopedFocus({
-				ancestorElement: scopedElement.querySelector('.slds-popover'),
+				ancestorElement: this.dialogContent,
 			}); // eslint-disable-line react/no-find-dom-node
 			// Don't steal focus from inner elements
 			if (!DOMElementFocus.hasOrAncestorHasFocus()) {
@@ -378,7 +398,11 @@ const Dialog = createReactClass({
 			this.props.position === 'absolute' ||
 			this.props.position === 'overflowBoundaryElement'
 		) {
-			style = this.getPopperStyles();
+			style = {
+				...style,
+				outline: 0,
+				...this.getPopperStyles(),
+			};
 		}
 
 		if (
@@ -399,9 +423,10 @@ const Dialog = createReactClass({
 				.getBoundingClientRect().width;
 		}
 
-		if (this.props.style) {
-			style = { ...style, ...this.props.style };
-		}
+		style = {
+			...style,
+			...this.props.style,
+		};
 
 		const contents = (
 			<div // eslint-disable-line jsx-a11y/no-static-element-interactions
@@ -414,6 +439,8 @@ const Dialog = createReactClass({
 							[`${this.props.outsideClickIgnoreClass}`]:
 								this.props.position === 'overflowBoundaryElement',
 						},
+						this.props.hasNubbin &&
+							getNubbinClassName(this.props.align, this.state.popperData),
 						this.props.contentsClassName
 					) || undefined
 				}
@@ -421,8 +448,10 @@ const Dialog = createReactClass({
 				onKeyDown={this.handleKeyDown}
 				onMouseEnter={this.props.onMouseEnter}
 				onMouseLeave={this.props.onMouseLeave}
-				{...this.props.containerProps}
 				ref={this.setDialogContent}
+				role={this.props.variant}
+				tabIndex={this.props.variant === 'popover' ? '-1' : undefined}
+				{...this.props.containerProps}
 			>
 				{this.props.children}
 			</div>